refactor(chart): one implementation per geom — Plot.* → geoms/*
Removes the duplicate prop-based Plot/{Bar,Line,Area,Point}.svelte primitives (thin
twins of the full geoms/* versions already used by every high-level chart). The Plot
namespace + Plot/index.js now point Bar/Line/Area/Point at geoms/* — matching how
Box/Violin/Jitter already worked. So the composable Plot.* gets the full-featured
geoms (grouped/stacked bars, symbols, selection, crossfilter, orientation flip) and
there's a single source per geom. Deleted the obsolete prop-based primitive specs +
Test helpers (geoms/* behaviour is covered by builder + chart + composable tests);
updated BarChartExample to the registration form.

BREAKING: Plot.Bar/Line/Area/Point are registration-based (<Plot.Root data><Plot.Bar
x y/></Plot.Root>) instead of prop-based (<Plot.Bar data=…/>), like Plot.Box.

Verified live: composable Plot.Bar renders 5 bars, per-category fills, correct axes,
0 NaN. Full suite 5319 green, svelte-check clean.
