+++ packages/chart/src/Plot/index.js
@@ -1,9 +1,9 @@
 export { default as Root } from './Root.svelte'
 export { default as Axis } from './Axis.svelte'
-export { default as Bar } from './Bar.svelte'
+export { default as Bar } from '../geoms/Bar.svelte'
 export { default as Legend } from './Legend.svelte'
 export { default as Grid } from './Grid.svelte'
-export { default as Line } from './Line.svelte'
-export { default as Area } from './Area.svelte'
-export { default as Point } from './Point.svelte'
+export { default as Line } from '../geoms/Line.svelte'
+export { default as Area } from '../geoms/Area.svelte'
+export { default as Point } from '../geoms/Point.svelte'
 export { default as Arc } from './Arc.svelte'
