+++ packages/chart/src/index.js
@@ -1,11 +1,11 @@
 import Root from './Plot/Root.svelte'
 import Axis from './Plot/Axis.svelte'
-import Bar from './Plot/Bar.svelte'
+import Bar from './geoms/Bar.svelte'
 import Grid from './Plot/Grid.svelte'
 import Legend from './Plot/Legend.svelte'
-import Line from './Plot/Line.svelte'
-import Area from './Plot/Area.svelte'
-import Point from './Plot/Point.svelte'
+import Line from './geoms/Line.svelte'
+import Area from './geoms/Area.svelte'
+import Point from './geoms/Point.svelte'
 import Arc from './Plot/Arc.svelte'
 import Highlight from './geoms/Highlight.svelte'
 import Trend from './geoms/Trend.svelte'
